[refactor] move get_{native,studied}_language to notes.language
app.utils is now empty
