+++ app/notes/__init__.py
@@ -1,2 +1,8 @@
-from .language import language_code_by_name, Language, get_language
+from .language import (
+    language_code_by_name,
+    Language,
+    get_language,
+    get_native_language,
+    get_studied_language,
+)
 from .note import Note, get_note
