+++ app/telegram/language.py
@@ -11,8 +11,12 @@ from nachricht.messenger import Context, Keyboard, Button
 from nachricht.i18n import TranslatableString as _
 
 from .. import bus, router
-from ..util import get_studied_language, get_native_language
-from ..notes import language_code_by_name, Language
+from ..notes import (
+    language_code_by_name,
+    Language,
+    get_studied_language,
+    get_native_language,
+)
 from ..srs import get_notes
 from .note import get_explanation_in_native_language
